move SerialPort to seperate module
